+++ dataland-community-manager/src/main/kotlin/org/dataland/datalandcommunitymanager/api/CompanyRolesApi.kt
@@ -8,7 +8,6 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
-import org.dataland.datalandbackendutils.model.KeycloakUserInfo
 import org.dataland.datalandbackendutils.utils.swaggerdocumentation.CommunityManagerOpenApiDescriptionsAndExamples
 import org.dataland.datalandbackendutils.utils.swaggerdocumentation.CompanyIdParameterRequired
 import org.dataland.datalandbackendutils.utils.swaggerdocumentation.CompanyRoleParameterRequired
@@ -295,44 +294,4 @@ interface CompanyRolesApi {
         )
         @PathVariable("companyId") companyId: UUID,
     )
-
-    /**
-     * Get all users whose email matches the suffixes defined in the company's emailSuffix attribute.
-     * @param companyId the company to check
-     * @return list of users with emails matching the suffixes
-     */
-    @Operation(
-        summary = "Get users by company email suffixes.",
-        description = "Returns all users whose email address ends with any of the suffixes defined in the company's emailSuffix attribute.",
-    )
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                responseCode = "200",
-                description = "Successfully retrieved users with matching email suffixes.",
-            ),
-            ApiResponse(
-                responseCode = "403",
-                description = "You do not have the right to make this query.",
-                content = [Content(array = ArraySchema())],
-            ),
-            ApiResponse(
-                responseCode = "404",
-                description = "The specified company does not exist or has no emailSuffix defined.",
-                content = [Content(array = ArraySchema())],
-            ),
-        ],
-    )
-    @GetMapping(
-        produces = ["application/json"],
-        value = ["/company-role-assignments/by-email-suffix/{companyId}"],
-    )
-    @PreAuthorize(
-        "hasRole('ROLE_ADMIN') or " +
-            "@SecurityUtilsService.isUserOwnerOrMemberAdminOfTheCompany(#companyId)",
-    )
-    fun getUsersByCompanyEmailSuffix(
-        @CompanyIdParameterRequired
-        @PathVariable("companyId") companyId: UUID,
-    ): ResponseEntity<List<KeycloakUserInfo>>
 }
